Fix e4 part stack styling and activation
The e4 CSS styling engine is not available in the RAP port. Style active
part stack manually in custom StylingEngineImpl class.

The Activate event in RAP does not have a "detail" flag set. Always
activate part stack regardless this flag.
