+++ bundles/org.eclipse.e4.ui.workbench.swt/build.properties
@@ -17,7 +17,8 @@ bin.includes = META-INF/,\
                icons/,\
                about.html,\
                plugin.properties
-source.. = src/
+source.. = src/,\
+           src-rap/
 src.includes = icons/,\
                about.html
 jre.compilation.profile = JavaSE-11
